server/requirements.txt
Beaker==1.6.4
PIL==1.1.7
amqp==1.0.11
anyjson==0.3.3
argparse==1.2.1
billiard==2.7.3.28
bottle==0.11.6
bottle-extras==0.1.0
bottle-redis==0.2.1
celery==3.1.10
celery-with-mongodb=3.0
distribute==0.6.35
gevent==0.13.8
gevent-websocket==0.3.6
greenlet==0.4.0
gunicorn==0.17.2
kombu==2.5.10
mongoengine==0.8.4
nose==1.2.1
pymongo==2.6.3
python-dateutil==2.1
python-memcached==1.48
pytz==2013b
redis==2.7.2
six==1.3.0
wsgiref==0.1.2
